ai-agents/requirements.txt
livekit-agents[deepgram,openai,cartesia,silero,turn-detector]~=1.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv
aiohttp~=3.10
imgkit~=1.2.3
Pillow>=10.3.0 